d/rules: allow external override of base dpdk configuration
By using := instead of = users that rebuild our packaging can overwrite
the following via environment variables:
  DPDK_CONFIG
  RTE_MACHINE
  RTE_TARGET

That is already used by some consumers of the Ubuntu packaging to get
optimized build, but would allow them to go on without a patch on top.
Also since it is a no-op, as long as the variables are not set, it is
fairly safe.

Change-Id: I26265c09c90d33a6e4cd37f8c26ac0ef5baac0c4
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>