d/rules: strip -Wl, from dpkg-buildflags output
The DPKG build system adds a -Wl, to and EXTRA_LDFLAGS, by that the
flags provided by dpkg-buildflags become -Wl,-Wl,option which fails to
build.
Fix by filtering the -Wl, prefix before exporting.

Note: filtering with make text functions is hard if the pattern contains
a comma, so sed was chosen.

Change-Id: I9b91b86c80529855dfb1c5a52725c0d07a95e5c8
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>