debian/*.dirs sanitize dir files v2
*updates in v2*
updated to adapt to changes we made when integrating earlier changes

This is a cleanup from the early days of dpdk packaging.
They are no more needed and not recommended if not needed according to
https://www.debian.org/doc/manuals/maint-guide/dother.en.html#dirs

The only remaining is a mkdir -p for dpdk_nic_bind which is much better
in d/rules close to the remove and relink that is done there.

Change-Id: Ife69e3fa23e5f061c86dec03d4b71d76a441e29d
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Signed-off-by: C.J. Collier <cjcollier@linuxfoundation.org>