librte-port3.symbols is now arch specific
since the following upstream commit librte-port3.symbols is arch
specific.
There was a feature added which itslef is arch neutral, but it is based on KNI
which is not 32 bit compatible, and thereby disabled on i386.
But the new feature of port to exploit it obviously depends on configured KNI,
therefore the arch difference in the symbols.

commit 9fc37d1c071c0b0227c0a41431d3beb2a14e31a3
Author: WeiJie Zhuang <zhuangwj@gmail.com>
Date:   Tue Jun 21 18:55:52 2016 +0800

    port: support KNI

Change-Id: I290ab11dfbeae0e301b5fe55d3404a9d46639c17
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>