scripts: make load-devel-config not to appear as executable
Quoting the first line of the script: "#! /bin/echo must be loaded with ."
Given that we should drop the .sh file ending as well as the executable
flag - both are not needed to source the file.

Backport of http://dpdk.org/dev/patchwork/patch/15100/
Already got an ack to be commited, but no git commit id yet.

Change-Id: Iadad8b2454feb82361731fce19928b395b256fa2
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>