build: enable powerpc
Enable ppc64el to build supported packages on that architecture as well.

xen dependencies are not available on ppc64el, so disable xen pmd in
that case.

powerpc currently has several libraries and pmds disabled in its default
dpdk config. That defines the range of expected-to-work packages.
Follow that and remove ppc64el architecture on those libs and pmds as packages.

Change-Id: I7f18962ad9e021c42e80739ddd569db4ecfe65dc
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>

Change-Id: Ic0bb0741f4027dd4ed371e3228aa770fb3b698f9
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>