add empty RTE_EAL_PMD_PATH dir to librte_eal
If EAL is initialized and the path isn't existing at all (=no PMDs at
all are installed) the initialization fails with:
 EAL: /usr/lib/x86_64-linux-gnu/dpdk-pmds/: cannot open shared object
 file: No such file or directory
So since we support RTE_EAL_PMD_PATH now but can never guarantee that
there always will be a PMD it becomes a requirement of librte_eal to
have at least an empty path.
An empty path is fine and will not load anything, but avoid the
initializaiton error.

Change-Id: I2725169c67c190f8ed8b608fc5ae508ae66eab1c
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>