add PMDs as recommend for dpdk runtime.
Once people pull in dpdk runtime they likely want to use it.
These days PMDs are usually not covered by depends (other libraries
are linked if used).
To avoid the issue of systems that would work if just those packages
would be installed recommend them.
It is no real "dependency" things can work, therefore no depends. But it
is as much recommends as recommendations can go.
This got a bit amplified by removing libdpdk-dev from the depends which
formerly pulled all in.

Change-Id: I03cc329412de9d37c4fd203d09c8e27812bc9e5b
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>