enable dpdk autotests as dep8 tests
The autotests we bundle can be used on machines able to run DPDK.
But sadly almost none of them can run inside most common build
environments which makes it unavailable for a classic "dh_autotest".
But we can define certain characteristics of an dep8 test.
In there at least the majority of the tests runs fine.

So this adds the autotests as dep8 test.
To do so it defines the test as "isolation-machine, needs-root", but
even then checks further required capabilities like sse3 and the
required huge pages.
If prereqs are not met it exits by skipping the test, which counts as
PASS.

The tests can run successful e.g. with an autopkgtest call that adds the
required CPU characteristics. To do so run it like usual but with:
  autopkgtest [...] -- qemu --qemu-options='-cpu qemu64,+ssse3' [...]

So far only amd64 is supported and as mentioned before only if the
prereqs are met. But it is written in a way that this can be the groundwork
for any other architecture that wants to enable the autotests as well.

Change-Id: Ia2a2839ce349b3ea5c22f0b8a589f8a71378adc1
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>