add new net library
The librte_net library is a collection of IP protocol definitions and
convenience macros. It is based on code from the FreeBSD* IP stack and
contains protocol numbers (for use in IP headers), IP-related macros,
IPv4/IPv6 header structures and TCP, UDP and SCTP header structures.

This is newly built by default as of DPDK 16.11, so make it available in
the packaging.

Change-Id: I7dff9347b54918d908a3ed436077470fcdef1969
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>