bump librte-eal version from 2 to 3 as package transition
This is a bit of a "decision" we don't have many (in Debian actually
none) depending packages so far.
In a library that would behave we would ship dpdk-16.11 as new source
along dpdk-16.07, but as discussed we don't want to do that just yet.
OTOH the version 3 is fully backward compatible with version 2, so for
now we can handle it like a package rename transition.

That implies to rebuild dependent packages in distributions that have
them already (Ubuntu) as a transitional package upload.

Change-Id: I92e79f206a0bd752d775584c88ea2559c652b606
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>