bump librte-cryptodev version from 1 to 2 as package rename transition
This is a bit of a "decision" we don't have many (in Debian actually
none) depending packages so far.
In a library that would behave we would ship dpdk-16.11 as new source
along dpdk-16.07, but as discussed we don't want to do that just yet.
OTOH the version 2 is fully backward compatible with version 1, so for
now we can handle it like a package rename transition.

That implies to rebuild dependent packages in distributions that have
them already (in Ubuntu) as a transitional joint package upload.

Change-Id: I028c24f6cc657731b1d3f058bd484fdb5ebacb49
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>