handle librte-ethdev version and rename changes package rename transition
This is a bit of a "decision" we don't have many (in Debian actually
none) depending packages so far.
In a library that would behave we would ship dpdk-16.11 as new source
along dpdk-16.07, but as discussed we don't want to do that just yet.
OTOH the version 5 is fully backward compatible with version 4, so for
now we can handle it like a package rename transition.

On top of the api bump the library got renamed by upstream from
libethdev to librte-ethdev to follow their usual naming scheme.

That implies to rebuild dependent packages in distributions that have
them already (in Ubuntu) as a transitional joint package upload.

Change-Id: I13e0fc99ac1adb230f1cfc7491eb7cf162996df3
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>