make test-autotest always pass
For now the autotest is too unreliable, so we run it to get some
logs in different environments, but never (want to ) fail until
it is stable.

This was already the intention and worked if certain tests failed
as we didn't check on the RC of the tests, but if the test overall
fatally fails we had a FAIL due to the set -e.

Change-Id: I54052351113a974d3e7810d55405f063e1fbe16a
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>