Undo library version transition packages
The old approach was just wrong, we have two cases:
1. a distribution without the old sonames, no transition needed
2. a distribution with the old sonames
   - continue to provide them by the old binary packages works fine
     (they are named libs for a reason)
   - provide new ones in new binary packages built now
   - whenever anything is rebuilt it picks up the new dependency

The worst is that I did this :-/ not thinking at all it seems.
Not hurtin on case #1, but a lot on case #2 - you likely want an
archive admin to drop the transitional packages after binaries with this
fix are built.

Change-Id: Ie60245818bc55a075a8b565dd38614351811e849
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>