debian/rules: adapt new location and build of tests
Build test tools which is now split from default build.
We want those as we already provided a few of them together with the dpdk-dev
package as helpers.
Since we want only those from the shared (not the static) build we only
add the shared ones.

Also along that some paths changed which had to be adapted.
This will add testacl and testpipeline tools and unify them with the
already provided test and testpmd tools in one directory.

Due to that testpmd will no more be in path, but that is correct given
that it is a test/dev tool more than a normal end user binary.

Change-Id: Ice383b5a70fe26fcd12864a0a86eebe114272108
Signed-off-by: Luca Boccassi <luca.boccassi@gmail.com>
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>