Hide dbgsym metapackage behind DEB_BUILD_OPTIONS
dbgsym packages go in the -debug section of the archive, and various
DDs and FTP Masters have strongly recommended to avoid uploading any
such metapackage to Main for this reason.
It can be still very useful for users and rebuilders, so leave it
disabled behind a dbgsym_meta DEB_BUILD_OPTIONS flag.

Change-Id: I112f3195755cc6bafef3167b39d3d47273572377
Signed-off-by: Luca Boccassi <luca.boccassi@gmail.com>