d/control: make python-pyelftools a recommends
It is a uncommon tool to use so it was never very important.
But pmdinfo needs the extra python modules to work properly.

$ dpdk-pmdinfo
Traceback (most recent call last):
  File "/usr/bin/dpdk-pmdinfo", line 14, in <module>
    from elftools.common.exceptions import ELFError
ImportError: No module named elftools.common.exceptions

We had it as a suggests so far as Ubunut would not have been able to
pick it up, but that is now accepted in [1].
Thanks for carrying that as suggests so far, but we could now enable it
as recommends for the next version of dpdk.

[1]: https://bugs.launchpad.net/ubuntu/+source/python-pyelftools/+bug/1630073

Change-Id: Ied23438a6332004f957e3e50842eeea98cfdac34
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>