d/rules: make auto-loaded drivers dir versioned
The dir registered as CONFIG_RTE_EAL_PMD_PATH= will be loaded by EAL.
But that is an issue if multiple versions of DPDK are co-installed.

Version A of EAL might load drivers of version B or vice versa.
Since this isn't the normal linker loads, but loads from a dir we must
version it to avoid issues of these cross-version loads.

Fixes: https://bugs.launchpad.net/ubuntu/+source/dpdk/+bug/1741244

Change-Id: I4e97f077f531f7fcb746a2c5b2ffec60d140a4a1
Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>