Fix the check for shared mbuf in rte_vlan_insert
An mbuf segment may have been cloned and the original mbuf segment
freed, in which case the mbuf segment is indirect but has a refcount
of 1 and it isn't shared. In this case, there is no need to allocate a
new mbuf segment since writing into the mbuf data won't cause any
adverse effects.

Therefore, fix the shared mbuf check in rte_vlan_insert to take into
account segments that are clones where the original mbuf has been
freed.

Fixes: VRVDR-49246