Remove debian "history"
The upstream debian repository hasn't preserved the upstream dpdk-stable
repository history. This doesn't matter since the build environment
uses the tagged version of the dpdk-stable tree when building. However,
it is nice to be able to exmaine the history while on this branch.
The next commit will merge the dpdk-stable history.