Mellanox 100G: Dataplane intf not up after admin down/up: remove trigger
The dataplane master thread becomes indefinitely blocked after admin
down. When checking if a port is a backplane intf on if_stop(),
rte_eth_dev_info_get() blocks due to the mlx5 driver waiting
indefinitely for the netlink msg to obtain the intf index from the
kernel. The problem is concurrent use of the same netlink socket from
multiple threads. The reason this only occurs on admin down is that
netlink messages are continuously sent and received on another thread
at an observed rate of over 4k/s due to the following call chain (i.e.
3 netlink msgs per link update):

mlx5_dev_interrupt_handler()
 - mlx5_dev_status_handler()
  - mlx5_link_update()
   - mlx5_link_update_unlocked_gs()
     - mlx5_ifreq()
      - mlx5_get_ifname()
       - mlx5_nl_ifindex()
      - mlx5_ifreq()
       - mlx5_get_ifname()
        - mlx5_nl_ifindex()
      - mlx5_ifreq()
       - mlx5_get_ifname()
        - mlx5_nl_ifindex()

The reason for this is the check in mlx5_link_update_unlocked_gs() by
which -EAGAIN is returned if link status does not correspond to link
speed. If status is DOWN, a ETH_SPEED_NUM_NONE (0) speed is expected.
However, as the link speed is -1 on admin down, -EAGAIN is returned.
The solution is to modify the check to account for this.

By getting rid of this trigger, the risk of an issue due to concurrent
use of the netlink socket is reduced for admin down to be the same as
on admin up.

VRVDR-48593