Mellanox 100G: Dataplane intf not up after admin down/up: fix concurrent netlink socket use
The dataplane master thread becomes indefinitely blocked after admin
down. When checking if a port is a backplane intf on if_stop(),
rte_eth_dev_info_get() blocks due to the mlx5 driver waiting
indefinitely for the netlink msg to obtain the intf index from the
kernel. The problem is concurrent use of the same netlink socket from
multiple threads. The reason this only occurs on admin down is that
netlink messages are continuously sent and received on another thread
at an observed rate of over 4k/s, this issue is addressed elsewhere.

The solution is to avoid having to use the netlink socket to retrieve
the intf index by caching this in the private data for the device.
This also has the benefit of improved performance given the prevalence
of calls to obtain the ifindex and/or ifname. Note that the problem has
been addressed in dpdk v19.08 by fa2e14d4921a ("net/mlx5: cache
associated network device index"):

"The associated device index is retrieved via Netlink request to
underlying Infiniband device driver. This network device index
is permanent throughout the lifetime of device. We do not
spawn the rte_eth_dev ports without associated network device, and
if network device is being unbound we get the remove notification
message and rte_eth_dev port is also detached. So, we may store
the ifindex in mlx5_device_spawn() routine at rte_eth_dev port
creation and initialization time and use the cached value further
instead of doing actual Netlink request."

The upstream fix was not used due to dependency on a previous commit
ad74bc619504 ("net/mlx5: support multiport IB device during probing"),
and multiport IB device support relies on several commits not present
in v18.11.

VRVDR-48593