net/mlx5: fix link speed info when link is down
One of the fixes for VRVDR-48593 corrects mlx5_link_update_unlocked_gs()
so that it does not return -EAGAIN when the link is down due to the
link speed being -1 instead of the expected ETH_SPEED_NUM_NONE (0).
While this fix works, an upstream fix added in v19.08 deals with this
issue but was not noticed at the time. In the interim period, the
upstream fix was backported to v18.11.5 as d72296ec9756 ("net/mlx5: fix
link speed info when link is down"). While the local patch is harmless,
remove it as it has been superseded by the upstream patch.

VRVDR-49431