net/virtio: use promiscuous mode when secondary MAC fails
The addition (and removal) of secondary MAC addresses may not be
supported by the underlying "backend" driver. When the port is being
used as part of a bond interface, the failure propagates through to
bonding resulting in a part-configured interface. See
slave_add_mac_addresses() and __eth_bond_slave_add_lock_free().

If there is no support for secondary MAC addresses, have the add
function turn on promiscuous mode. Similarly on removal of the last
secondary address, turn off promiscuous mode.

Fixes: VRVDR-51067