net/mlx5: fix out-of-bound write in xstats
One way to determine the number of extended statisitics available is to
query to a length of 0. However, when xstats is not yet initialized,
this passes the initial check aginst the known mlx5 statistics (also
zero at this point). Later, when the counters are finally read, they
are written back to the temporary stack copy that was allocated with
a length of 0.

To fix this, pass down to mlx5_read_dev_counters what is really
available for writing.

Bugfix: VRVDR-46357