net/i40e: disable source pruning
By default, the i40e does source pruning so it doesn't see its own
packets. That is desireable for some configurations. For other
protocols, like LAG and VRRP, this may become an issue if you add
additional MAC addresses that are shared with other network devices.
When this happens, the device needs to see its "own" packets coming
from the network.

Bugfix: VRVDR-52611