Fix performance issue when shaping to line rate
VRVDR-52877

When we allow borrowing of tokens from the shaper bucket we will move
the port time forward by the packet size in grinder_schedule().
This resulted in the number of periods being calcaulted in
grinder_credits_update() being more than they should have been resulting
in large bursts of traffic.  To avoid this we now use the field
time_cpu_bytes to calculate the number of periods since this will reflect
the real time and not be wound forward like port->time.