Exit the dequeue loop if all pipes have been checked and there's no tokens
VRVDR-53324

Return the token state when they're check in grinder_schedule() so we can
check in the main dequeue loop whether all pipes have been checked and
none have any tokens left.  If this is the case we can exit the loop
because there's no more packets to be dequeued.