Make sure the calculated rational numbers aren't too high.
VRVDR-53324

When the shaper period and credit values are calculated if the accuracy
is too high the result is large values can be returned.  This leads to
a bursty traffic profile.  To try to minimize this reduce the accuracy
so lower number pairs will be returned.

(cherry picked from commit 44a8f2666a84e0e4525c668c40d2ab5b309878a9)