Make sure the calculated rational numbers aren't too high.
VRVDR-53324

When the shaper period and credit values are calculated if the accuracy
is too high the result is large values can be returned.  This leads to
a bursty traffic profile.  To try to minimize this reduce the accuracy
so lower number pairs will be returned.

(cherry pick from git 7f33bf96f8a93c3376b0e1fd02323fbf8f938e87)