rte-acl: add rte_acl_del_rule API
This allows removal of individual rules of an existing rte_acl_ctx,
but requires, as for rte_acl_ad_rules an rte_acl_build invocation.

This got introduced to meet the requirements for a more dynamic/runtime
driven ACL usage.

This allows an alternative rte_acl_rule storage inside
librte_acl based on rte_hash and rte_mempool. The usage of the
hashtable/mempool is optional in general. But mandatory if the
rte_acl_del API needs to be used.

The usage of the hashtable/mempool implementation inside rte_acl
gets enabled via the rte_acl_param flag ACL_F_USE_HASHTABLE.

VRVDR-53713