net/bonding: fix potential endless loop on closing
If stop fails, we skip it and also continue to attempt to remove the
device from the bonding group. This removal will likely fail but you
should only remove stopped device and a failure to stop indicates a
deeper problem. This second failure increments skipped again and we
are now past the check in the while.

To avoid this issue, do not attempt any further cleanup if we can't
stop the device.

Fixes: fb0379bc5db3 ("net/bonding: check stop call status")
Cc: stable@dpdk.org

Bugfix: VRVDR-54505